+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.13.1 - 2023-12-26
+- Upgrade ampltools to 0.7.2 adding amplpy.bundle command.
+- Fix issue with environment variables on Windows in a generic way.
+
 ## 0.13.0 - 2023-12-26
-- Upgrade ampltools to 0.7.1 adding amplpy.bundle command.
 - [Breaking] Modules now have priority in the PATH.
 - Fix ampl_libpath issue on Windows.